Refactor/#50: app.js 파일 위치 변경 및 로직 변경 (app.js - Express 설정 / server.js - DB 연결+서버 실행)
